chore: move app_sizes.dart to the right place
