+++ lib/features/debug/presentation/ui_kit/ui_kit_wm.dart
@@ -1,8 +1,8 @@
 import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_template/common/mixin/localization_mixin.dart';
-import 'package:flutter_template/features/app/di/app_scope.dart';
 import 'package:flutter_template/common/mixin/theme_wm_mixin.dart';
+import 'package:flutter_template/features/app/di/app_scope.dart';
 import 'package:flutter_template/features/debug/presentation/ui_kit/ui_kit_model.dart';
 import 'package:flutter_template/features/debug/presentation/ui_kit/ui_kit_screen.dart';
 import 'package:flutter_template/features/theme_mode/presentation/theme_mode_provider.dart';
