+++ lib/uikit/themes/app_theme_data.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_template/features/common/utils/sizes/app_sizes.dart';
+import 'package:flutter_template/uikit/app_sizes.dart';
 
 import 'package:flutter_template/uikit/colors/app_color_scheme.dart';
 import 'package:flutter_template/uikit/text/app_text_scheme.dart';
